+++ package.json
@@ -9,6 +9,7 @@
     "axios": "^1.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.12.1",
     "react-scripts": "^5.0.1",
     "react-transition-group": "^4.4.5",
     "web-vitals": "^2.1.4"
