+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default Footer;
