+++ src/components/Footer/MyFooter.module.css
@@ -1,0 +1,3 @@
+.myFooter{
+
+}
