+++ src/components/Header/MyHeader.module.css
@@ -1,0 +1,12 @@
+.myHeader{
+    height: 50px;
+    width: 100vw;
+    display: flex;
+    align-items: center;
+    padding: 0 15px;
+    background: pink;
+}
+
+.myHeaderNav{
+    margin-left: auto;
+}
