+++ src/styles/App.css
@@ -15,7 +15,8 @@ hr{
 
 #root{
     display: flex;
-    justify-content: center;
+    flex-direction: column;
+    align-items: center;
 }
 
 .App{
@@ -96,6 +97,5 @@ div > form{
     display: flex;
     flex-direction: column;
 }
-
